[rename] create utils directory, move to file - Move queryKeys.ts to /utils Github issue #37
